First version with full legal move (tested with PERFT and corrected). Adding basic CLI to explore data.

+++ main.cpp
@@ -1,17 +1,32 @@
 #include "Game.h"
+#include "CLI.h"
 
 int main(int argc, char** args) {
-	Game *pGame = new Game();
+	// If argument, enter test mode
+	if (argc > 1) {
+		if (strcmp(args[1], "-g") == 0 || strcmp(args[1], "-graphic") == 0) {
+			Game* pGame = new Game();
 
-	// Initializing the game. If failed end the program with error code.
-	if (!pGame->init()) return 1;
+			// Initializing the game. If failed end the program with error code.
+			if (!pGame->init()) return 1;
 
-	pGame->run();
+			pGame->run();
 
-	// Freeing memory
-	pGame->kill();
-	delete pGame;
+			// Freeing memory
+			pGame->kill();
+			delete pGame;
 
-	// Ending program
+			return 0;
+		}
+	}
+
+	CLI cli;
+	
+	cli.init();
+
+	cli.run();
+
+	cli.kill();
+
 	return 0;
 }
